match gpt-neox structure

+++ gpt_k/model/modules/positional_encoding.py
@@ -14,10 +14,10 @@ class PositionalEncoding(nn.Module):
     def __init__(self, dropout, seq_len, embedding_dim):
         super().__init__()
         self.P = torch.zeros([1, seq_len, embedding_dim])
-        X = torch.arange(seq_len, dtype=torch.float32) # vector of indices up to seq_len
+        X = torch.arange(seq_len) # vector of indices up to seq_len
         X = X.view(-1, 1) # unsqueeze all elements of the vector individually [seq_len, 1]
         # part one of the positional encoding, batched across the sequence length rather than looped
-        untitled_var = torch.pow(10000, torch.arange(0, embedding_dim, 2, dtype=torch.float32) / embedding_dim) # [512]
+        untitled_var = torch.pow(10000, torch.arange(0, embedding_dim, 2) / embedding_dim) # [512]
         X = X / untitled_var # [seq_len, pe_len]
         # part two of the positional encoding
         self.P[:, :, 0::2] = torch.sin(X) # sin alternating from index 0 (even)
